Add option for stricter compiler warnings (#1891)
- Add cmake-option to use a stricter compiler-warning level, default OFF
- Apply stricter compiler warnings for gfxrecon_format|util|graphics targets

+++ tools/tocpp/main.cpp
@@ -364,7 +364,7 @@ bool ProcessCapture(gfxrecon::decode::VulkanCppConsumer&      cpp_consumer,
 
     do
     {
-        printf("  Processing Frame %u\r", file_processor.GetCurrentFrameNumber());
+        printf("  Processing Frame %u\r", static_cast<uint32_t>(file_processor.GetCurrentFrameNumber()));
         fflush(stdout);
         success = file_processor.ProcessNextFrame();
     } while (success && file_processor.GetCurrentFrameNumber() <= frame_limit);
